% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energystar.R
\name{estar_list}
\alias{estar_list}
\title{List all known ENERGY STAR product categories}
\usage{
estar_list()
}
\value{
tibble with columns: category, view_id, description
}
\description{
Returns a tibble of ~70 product categories with view_id for use in
estar_products() or estar_view().
}
