% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energystar.R
\name{estar_search}
\alias{estar_search}
\title{Search ENERGY STAR product categories by keyword}
\usage{
estar_search(query)
}
\arguments{
\item{query}{Character string to match against category names and descriptions}
}
\value{
tibble of matching categories
}
\description{
Search ENERGY STAR product categories by keyword
}
